OPENDJ-2354 Backport OPENDJ-2247: CertAndKeyGen class location is incorrect when running on IBM Java 8
This is a small fix/cleanup to Platform.java CertAndKeyGen handling and add support for Java 8.

The various if/then/else constructs were starting to get out of hand so I changed this to just check a list of known/possible locations.

There is also a java property introduced that could allow it to be easily overridden in the future without a patch.